moving reads to external drive MinIT
Script to be run from cron.hourly oin your MinIT  to move folders older than an hour to external storage  from /data.
It moves files in a run folder to an external drive, checks if they are there, then deletes them if they are identical on both the external drive and on /data. Prerequisite is that every run folder should start with run_ if you want them to be processed by the script.  Please use at your risk. I'm not responsible for file loss or anything that happens when running this script.